+++ src/data_gradients/batch_processors/output_mapper/tensor_extractor.py
