+++ src/data_gradients/batch_processors/segmentation.py
@@ -1,7 +1,7 @@
 from typing import List
 
 from data_gradients.batch_processors.base import BatchProcessor
-from data_gradients.batch_processors.adapters.dataset_adapter import DatasetAdapter
+from data_gradients.batch_processors.output_mapper.dataset_output_mapper import DatasetOutputMapper
 from data_gradients.batch_processors.preprocessors.segmentation import SegmentationBatchPreprocessor
 from data_gradients.batch_processors.formatters.segmentation import SegmentationBatchFormatter
 from data_gradients.config.data.data_config import SegmentationDataConfig
@@ -18,7 +18,7 @@ class SegmentationBatchProcessor(BatchProcessor):
         threshold_value: float = 0.5,
     ):
 
-        dataset_adapter = DatasetAdapter(data_config=data_config)
+        dataset_output_mapper = DatasetOutputMapper(data_config=data_config)
         formatter = SegmentationBatchFormatter(
             class_names=class_names,
             class_names_to_use=class_names_to_use,
@@ -27,4 +27,4 @@ class SegmentationBatchProcessor(BatchProcessor):
         )
         preprocessor = SegmentationBatchPreprocessor(class_names=class_names)
 
-        super().__init__(dataset_adapter=dataset_adapter, batch_formatter=formatter, batch_preprocessor=preprocessor)
+        super().__init__(dataset_output_mapper=dataset_output_mapper, batch_formatter=formatter, batch_preprocessor=preprocessor)
